requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.36
asyncpg==0.30.0
redis[hiredis]==5.2.1
pydantic-settings==2.8.1
httpx==0.28.1
python-dotenv==1.0.1

greenlet==3.1.1
gspread==6.2.1
google-auth==2.40.3
